# docs/openapi.yaml
openapi: 3.0.3
info:
  title: mStream API
  version: "1"
  description: |
    REST API for the [mStream](https://github.com/IrosTheBeggar/mStream) music streaming server.

    ## Authentication

    Most endpoints require a JWT bearer token. Tokens are issued by `POST /api/v1/auth/login` and
    can be supplied via any of (checked in order):

    1. Request body field `token`
    2. Query parameter `token`
    3. HTTP header `x-access-token`
    4. Cookie `x-access-token`

    ### Public mode

    If no users exist in the database, all requests are authenticated as a virtual public user with
    full library access — no token required. This is the default for fresh installs.

    ### Special tokens

    - **Admin tokens** have `admin: true` in the payload and are required for every `/api/v1/admin/*` endpoint.
    - **Jukebox tokens** (`jukebox: true`) are short-lived and bound to an active remote-control session.
    - **Shared-playlist tokens** (`shareToken: true`) restrict file access to the tracks in one specific shared playlist.
    - **Federation invite tokens** (`invite: true`) can only call `POST /api/v1/federation/invite/exchange`.

    ## Errors

    Errors use the shape `{ "error": "<message>" }`. Status codes:

    | Code | Meaning |
    |------|---------|
    | 400  | Bad request (invalid input) |
    | 401  | Unauthenticated |
    | 403  | Forbidden (Joi validation error, missing permission) |
    | 404  | Resource not found |
    | 405  | Admin API locked |
...

...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Pass the JWT in the `x-access-token` header. Despite the OpenAPI "bearer" label,
        the value does NOT include a `Bearer ` prefix — it's the raw token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: x-access-token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
      example:
        error: Authentication Error
    Empty:
      type: object
      description: Empty success response — the operation completed with no additional data.
      example: {}
...
  responses:
    EmptySuccess:
      description: Operation succeeded.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Empty" }
    Unauthorized:
      description: Auth required or token invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Validation error or permission denied.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    AdminLocked:
      description: Admin API disabled (either by configuration or because the user is not an admin).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/v1/file-explorer:
    post:
      tags: [File Explorer]
      summary: Browse a directory within the user's libraries
      description: Pass an empty string or `~` to list vpaths / home.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [directory]
              properties:
                directory:
                  type: string
                sort:
                  type: boolean
                  default: true
                pullMetadata:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Directory listing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string }
                  directories: { type: array, items: { type: object } }
                  files: { type: array, items: { type: object } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/file-explorer/recursive:
    post:
      tags: [File Explorer]
      summary: Recursively list all audio files under a directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [directory]
              properties:
                directory: { type: string }
      responses:
        "200":
          description: Array of virtual paths.
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/file-explorer/mkdir:
    post:
      tags: [File Explorer]
      summary: Create a directory
      description: Requires per-user mkdir permission. Recursive.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [directory]
              properties:
                directory: { type: string }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/file-explorer/upload:
    post:
      tags: [File Explorer]
      summary: Upload a file
      description: |
        Multipart form upload via busboy. Target directory goes in the `data-location` header.
        Requires per-user upload permission. Filenames are sanitized to prevent path traversal.
      parameters:
        - name: data-location
          in: header
          required: true
          schema: { type: string }
          description: Target directory as virtual path.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/file-explorer/m3u:
    post:
      tags: [File Explorer]
      summary: Parse an M3U playlist file
      description: Returns resolved paths for each entry, sanitized against directory traversal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string }
      responses:
        "200":
          description: Parsed entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        type: { type: string }
                        name: { type: string }
                        path: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ── Download ─────────────────────────────────────────────────────────────
...
  /api/v1/db/waveform:
    get:
      tags: [Waveform]
      summary: Get waveform peaks for a track
      description: 800 bars, 0–255 each. Generated on demand and cached.
      parameters:
        - name: filepath
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Waveform data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  waveform:
                    type: array
                    items: { type: integer, minimum: 0, maximum: 255 }
                    minItems: 800
                    maxItems: 800
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ── Lyrics ──────────────────────────────────────────────────────────────
...
  /api/v1/admin/file-explorer:
    post:
      tags: [Admin - File Explorer]
      summary: Browse the entire filesystem (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [directory]
              properties:
                directory: { type: string, description: "Absolute path, or `~` for home." }
                joinDirectory: { type: string }
      responses:
        "200":
          description: Listing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string }
                  directories: { type: array, items: { type: object } }
                  files: { type: array, items: { type: object } }
        "405": { $ref: "#/components/responses/AdminLocked" }
...